+++ src/data/schema/brands.ts
@@ -1,4 +1,4 @@
-import { pgTable, serial, varchar, numeric } from 'drizzle-orm/pg-core';
+import { pgTable, serial, varchar, integer } from 'drizzle-orm/pg-core';
 import { relations } from 'drizzle-orm';
 import { countries } from './countries';
 import { timestamps } from './common/columns.helpers';
@@ -7,7 +7,7 @@ export const brands = pgTable('brands', {
     id: serial('id').primaryKey(),
     name: varchar('name', { length: 255 }).notNull(),
 
-    countryId: numeric('country_id').references(() => countries.id),
+    countryId: integer('country_id').references(() => countries.id),
     ...timestamps,
 });
